% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes-ECM.R
\name{plot.BayesECMpred}
\alias{plot.BayesECMpred}
\title{Plot the results from Bayesian ECM}
\usage{
\method{plot}{BayesECMpred}(
  x,
  index = 1,
  colpal = "Zissou1",
  cat_labels = rownames(x$Zdist[[1]]),
  features = c("Expectation line", "Expectation text"),
  ...
)
}
\arguments{
\item{x}{an object of \code{class} \code{"BayesECMpred"} obtained as the trained model output from the \code{\link[=predict.BayesECM]{predict.BayesECM()}} function.}

\item{index}{integer corresponding to the row number of \code{Ytilde} provided to \code{\link[=predict.BayesECM]{predict.BayesECM()}} to be plotted.}

\item{colpal}{color palette provided to \code{\link[grDevices:palettes]{grDevices::hcl.colors()}} to be used for plotting.}

\item{cat_labels}{character vector used on the plot legend to specify each category.  Note: the order of the categories provided in this string must match \code{rownames(x[[1]])} for the resulting plot to be accurate.}

\item{features}{character vector specifying ploting features to include.  \code{"Expectation line"} plots a vertical line located at the expectation of each category distribution.  \code{"Expectation text"} prints the expected value of each category distribution.  Setting \code{features = NULL} will remove the labels and lines for the expectations of the distributions.}

\item{...}{additional arguments passed to the generic \code{\link[base:plot]{base::plot()}} function.}
}
\value{
Plot showing densities and expectations for each event category for the stipulated event.
}
\description{
Plots predictive densities of category probabilities event is from each category of interest obtained from the \code{\link[=predict.BayesECM]{predict.BayesECM()}} function.
}
